Adds longhorn-share-manager v1.7.0 rock
The build process differs a bit from the original Dockerfiles, this
mainly has to do with how the building process is done through
rockcraft. Mainly, we need nfs-ganesha to be linked to its .so file.
This is typically done by running ldconfig in the final image, however
we can't do that. Thus, we run it in the build stage and copy the
generated ld.so.cache in the final image.

Added sanity test for the mentioned rock.