Adds longhorn-share-manager v1.6.2 rock (#22)
Based on the v1.7.0 rock and upstream Dockerfile. The golang version was
updated. The nfs-ganesha version V5.7. Does not include the libjson-c* package.

We're applying a patch to the downloaded ntirpc v5.0 in order to be able
to build it. This building issue has been resolved in v5.8 and newer.

Updates unit test to also test the new image.