Run integration tests on multipass
There are a few issues running the Longhorn integration tests on an LXD
runner, as we currently are, mainly having to do with ISCSI volumes
being exposed and mounted in LXD.

Running on host has its own issues, mainly the Pods have cluster-wide
connectivity issues.

Instead, we'll be testing Longhorn in a multipass VM. We also remove
some of the GitHub runner's bloat in order to free up sufficient space
for the VM and test. Note that according to the Longhorn deployment
guide, Longhorn requires 25% of the root disk to be available (if no
other disks are available).