Fixes longhorn-manager arm64 rocks
The ``iputils`` snap does nto have an arm64 target, so we can't use that
for arm64.

The iputils rpm  from suse (installed in the original Dockerfile) contains the
following binaries: arping, clockdiff, ping, tracepath.
It also contains symlinks for ping6 and tracepath6.

We'll be installing equivalent apt packages for those binaries to keep
parity.